backend/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter>=25.2


celery==5.6.2
redis==7.1.1

psycopg2-binary==2.9.11

gunicorn==24.1.1

openai==2.16.0
ollama==0.6.1

#boto3==1.34.0   # keep only if you use S3

requests==2.32.5  # keep only if you make external HTTP calls

python-dotenv==1.2.1  # optional if not using django-environ


email-validator

Jinja2>=3.1.6
